feat: check if step is done

+++ postinstall/scripts/12-chrome.sh
@@ -4,11 +4,15 @@ set -e
 
 source ./scripts/helpers/status.sh
 
-echo -e "\nInstalling Google Chrome...\n"
+echo -e "\nInstalling Google Chrome..."
+
+if command -v google-chrome >/dev/null 2>&1; then
+    echo "Google Chrome is already installed. Skipping..."
+    exit 0
+fi
 
 trap 'rm -rf *.rpm' EXIT
 
-echo "Installing Google Chrome..."
 wget https://dl.google.com/linux/direct/google-chrome-stable_current_x86_64.rpm
 sudo dnf install ./google-chrome-stable_current_x86_64.rpm -y
